Basic GOAP Implemented
I ended up following a different tutorial (link here: https://www.youtube.com/watch?v=Q7aHXn_LypI&t=936s) as my first tutorial that I tried to follow wouldn't cooperate and I needed a simpler solution than blindly poking around until something fixes itself. This new tutorial also has the benefit of being far simpler to understand as there are no factory or builders, everything is nice and seperated over scripts and is somewhat simpler to set up new goals or actions. One thing to note is that the prerequisites are baked into being checked in a function so might not allow for chaining goals together. This will inevitebly mean that the scripts responsible for doing more complex actions are going to be far bigger and more complicated. The next system to implement (and has the variable set up at least) will be health and stamina that depletes as the AI moves, the AI will then need to visit specific locations to start to regain the specified stat