Started Adding Stat System
Started adding a stat system with health and stamina that drop as you move tiles. There are set spots around the grid for where these can be replenished. This proved to be far more confusing and time consuming than anticipated but I remain optimistic (I am yet to test the system after going back to the roots of the system and removing my overcomplications) that this shouldn't take too much longer.
It should be noted that I have made use of the goto function as I really can't think of any other way to handle the wonky logic of saying whether a node needs to be repicked. The thing that could improve it iis using an enum as the return instead of an int but it works so I'm not touching it right now.
The next system to implement is testing and ironing out how having mutliple AIs will work. My thoughts are that it will need another grid per AI on scene as each AI has there own nodes that they have visited, version, etc.