Added functionality for multiple AIs and further refinement
Added support for multiple AI in the form of having to spawn a grid per AI but the goals and actions are perfectly modular between AIs so that was no issue. Had a slight issue with AIs getting stuck on the stat refill areaa as they would drop to urgent, fill up enough to lower their priority bellow wandering and then have to immediatly come back after taking a small step. Fixed this by complety filling up their stat.